refactor: merge files by role in domain
