+++ internal/http/handler/helpers_test.go
@@ -7,15 +7,15 @@ import (
 )
 
 type MockAuth struct {
-	RegisterFunc func(ctx context.Context, email domain.Email, password domain.Password) error
-	LoginFunc    func(ctx context.Context, email domain.Email, password domain.Password) (string, error)
+	RegisterFunc func(ctx context.Context, email domain.Email, password domain.UserPassword) error
+	LoginFunc    func(ctx context.Context, email domain.Email, password domain.UserPassword) (string, error)
 }
 
-func (m *MockAuth) Register(ctx context.Context, email domain.Email, password domain.Password) error {
+func (m *MockAuth) Register(ctx context.Context, email domain.Email, password domain.UserPassword) error {
 	return m.RegisterFunc(ctx, email, password)
 }
 
-func (m *MockAuth) Login(ctx context.Context, email domain.Email, password domain.Password) (string, error) {
+func (m *MockAuth) Login(ctx context.Context, email domain.Email, password domain.UserPassword) (string, error) {
 	return m.LoginFunc(ctx, email, password)
 }
